refactor(settings): extract settings file persistence into extension
Lifts the JSON settings-file load/save/observe code (~218 lines) out of
BellithSettings.swift into BellithSettings+Persistence.swift via an
extension. Covers loadSettingsFileIfNeeded, applySettingsFileData,
applyPersistedValue, persistSettingsFileIfNeeded, settingsFileData,
roundedForSettingsFile, startObservingSettingsFileIfNeeded,
reloadSettingsFileFromDiskIfNeeded, plus the two featureFlags helpers
they need.

Five access bumps from private to internal: PersistedKeys, settingsFileURL,
settingsFileObserver, lastPersistedSettingsFileData, storedFeatureFlags.

BellithSettings.swift: 1142 → 924 lines (-19.1%).
